♻️ restructure code
