+++ shared.py
@@ -1,0 +1,3 @@
+from fastmcp import FastMCP
+
+mcp = FastMCP("deepauto intern assignment mcp server")
