+++ tools/greeting.py
@@ -1,0 +1,8 @@
+from shared import mcp
+
+@mcp.tool
+def greeting(name: str) -> str:
+    """
+    Greet the user with name!
+    """
+    return f"Hello {name}!"
